delete FolloweringItem, FollowingList and related files
